fix: refactor assert funcs
