+++ tests/test_transpiler_unroller.py
@@ -39,36 +39,6 @@ def test_translate_gate_error_2q():
         translate_gate(gates.CZ(0, 1), natives)
 
 
-def test_assert_decomposition():
-    circuit = Circuit(2)
-    circuit.add(gates.CZ(0, 1))
-    circuit.add(gates.Z(0))
-    circuit.add(gates.M(1))
-    assert_decomposition(circuit, native_gates=NativeGates.default())
-
-
-def test_assert_decomposition_fail_1q():
-    circuit = Circuit(1)
-    circuit.add(gates.X(0))
-    with pytest.raises(DecompositionError):
-        assert_decomposition(circuit, native_gates=NativeGates.default())
-
-
-@pytest.mark.parametrize("gate", [gates.CNOT(0, 1), gates.iSWAP(0, 1)])
-def test_assert_decomposition_fail_2q(gate):
-    circuit = Circuit(2)
-    circuit.add(gate)
-    with pytest.raises(DecompositionError):
-        assert_decomposition(circuit, native_gates=NativeGates.default())
-
-
-def test_assert_decomposition_fail_3q():
-    circuit = Circuit(3)
-    circuit.add(gates.TOFFOLI(0, 1, 2))
-    with pytest.raises(DecompositionError):
-        assert_decomposition(circuit, native_gates=NativeGates.default())
-
-
 @pytest.mark.parametrize(
     "natives_2q",
     [NativeGates.CZ, NativeGates.iSWAP, NativeGates.CZ | NativeGates.iSWAP],
